Transaction Repository has been added and its model has been moified
Charge driver user view has been completed and debugged

+++ ParkingControlWeb/Helpers/Helper.cs
@@ -12,7 +12,7 @@ namespace ParkingControlWeb.Helpers
         static AesCryptography aesCryptography = new AesCryptography();
         static LocalDate localDate = new LocalDate();
 
-        public static float CalculateExpense(Expense expense) => expense.Calculate();
+        public static int CalculateExpense(Expense expense) => expense.Calculate();
 
         public static string DateShow(DateTime date)
         {
@@ -47,13 +47,13 @@ namespace ParkingControlWeb.Helpers
     public class Expense
     {
 
-        public float expense;
+        public int expense;
 
-        public Expense(float enterPrice, float hourlyPrice, float dailyPrice, float minutesSpent,float hoursSpent, float daysSpent)
+        public Expense(int enterPrice, int hourlyPrice, int dailyPrice, int minutesSpent, int hoursSpent, int daysSpent)
         {
             expense += enterPrice;
         }
 
-        public float Calculate() => expense;
+        public int Calculate() => expense;
     }
 }
